Renamed colors package to follow american style spelling to ensure CLI is consistent; rename doStuff function into printHelpCostum; renamed tty dependency into isatty
